feat: add missing argument when calling removeList functio in addLitenersToLists.js; move createNewList function to separate file
